requirements.txt
PyQt5>=5.15
numpy-stl>=3.0
numpy>=1.24
